+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add debug option
+- Add additional test coverage
+
 ### Changed
 
 - Don't throw CI error for this repository