+++ CHANGELOG.md
@@ -11,11 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add debug option
 - Add additional test coverage
-- Add badges to README
 
 ### Changed
 
 - Don't throw CI error for this repository
+- Cleanup README
 
 ## [0.1.1] - 2019-08-01